# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation Laravel App Abdul Goffar",
        "version": "1.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "description": "get user data by token auth",
                "operationId": "get data user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "id",
                                            "type": "integer",
                                            "example": "123"
                                        },
                                        "name": {
                                            "description": "name",
                                            "type": "string",
                                            "example": "abd goffar"
                                        },
                                        "email": {
                                            "description": "email",
                                            "type": "string",
                                            "example": "abd@gmail.com"
                                        },
                                        "password": {
                                            "description": "password",
                                            "type": "string",
                                            "example": "abd goffar 12345"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuthU": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "user"
                ],
                "description": "user registration",
                "operationId": "post data user",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name",
                                        "type": "string",
                                        "example": "abd goffar"
                                    },
                                    "email": {
                                        "description": "email",
                                        "type": "string",
                                        "example": "abd@gmail.com"
                                    },
                                    "password": {
                                        "description": "password",
                                        "type": "string",
                                        "example": "abd goffar 12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "description": "name",
                                            "type": "string",
                                            "example": "abd goffar"
                                        },
                                        "email": {
                                            "description": "email",
                                            "type": "string",
                                            "example": "abd@gmail.com"
                                        },
                                        "password": {
                                            "description": "password",
                                            "type": "string",
                                            "example": "abd goffar 12345"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Data Invalid"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuthC": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuthP": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuthU": {
                "type": "http",
                "description": " Use /auth to get the JWT token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}